requirements.txt
numpy
opencv-python
pandas
Pillow
pytorch
scikit-learn
scipy
seaborn
torchvision
matplotlib
tqdm